[breaking] Enable tests in TypeScript packages by default
In addition to install and build we'll now run the test command, unless
`dontTest` is set to true. The test command defaults to `yarn test`, but
can be changed in the package config (much like install and test).

This is a breaking change as TypeScript packages may not build anymore
because of broken tests.