[breaking] Introduce remote-pull cache level for remote download only
Prior to this change there was `remote-push` only. It's semantic was
confusing and it was implemented as `remote-pull`. This makes this a
breaking change as `remote-push` behaves differently now. Now the
two cache level are as follows:
- `remote-pull` allows remote download only, but no upload
- `remote-push` allows remote upload only, but no download

The latter is handy to fix cache-poisoning issues, the former is handy
to prevent them.