[bugfix] Fix ephemeral package dependencies
Ephemeral packages could cause the build to break due its dependencies
possibly not being built. Ephemeral packages do not fail TypeScript pkgs
if their transitive dependencies aren't built. Together with the local
cache this can lead to PkgNotBuiltErr. For example:
- `A` is a TypeScript library
- `B` is an ephemeral generic package
- `C` is a TypeScript offline mirror
`A` depends on `B` and  `B` depends on `C`. If ephemeral packages were
cached, i.e. `B` would not always be built, the build for `A` would
break if `C` isn't built yet, because `B` is ephemeral.

This change makes it so that ephemeral packages always need to be built,
but never break when they're not built as part of a transitive dependency.
This way they're just built when needed, but also aren't locally cached
across builds.