Go: Ignore ephemeral packages when checking transitive deps
Consider the case where you have three Leeway Go packages A, B, C where A depends on B. B depends on C and C is ephemeral.

The ephemeral package C will always be built whenever B is built as it is never cached.

However, if B doesn't need to be rebuilt, then C won't be built either.

As it is ephemeral it won't be in the cache either, which means that if you try to build package A when B is cached, then the build will fail as C isn't in the cache.

This commit fixes it so that ephemeral packages are ignored when checking the transitive dependencies of Go packages.

This was already the case for Yarn packages - the implementation (and comments) have been copied over verbatim