Add docker image results to the HTML report
This also changes the structure of the report so that it starts with a <table> that provides a quick overview of the packages that were built. Each how also contains a link to further down in the report as each package now has a separate <h2> section with details about the package. This makes it easier to add more details as we go without loosing the nice consice overview