Only use buildContext when building packages
pkg/build.Build may be called with a BuildOptions that contains a
pre-generated BuildContext; in this case the BuildOptions value is
largely ignored during context setup and may not be complete. The
buildContext struct will be authoritative in this context and must be
be used instead of buildOptions.

One way that this issue manifests is when running a leeway script with
dependencies as `leeway run` generates a buildContext and passes that
into leeway.Build; the rest of the buildOptions is empty. When using
buildOptions the RemoteCache field will always be empty, even if the
buildContext value has a RemoteCache defined. By only using buildContext
instead of buildOptions we resolve this confusion on what data is authoritative.